+++ src/features/movieList/FilterByYear.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import useDebounce from "../../../utilities/UseDebounce.jsx"
+import useDebounce from "../../utilities/UseDebounce.jsx"
 
 
 const FilterByYear = ({ value, onChange }) => {
